Working! Sends and receives messages to remote nodes

+++ src/nodes/Master.cpp
@@ -8,7 +8,6 @@
 
 // Constructor: initializes the apprentice list and reads the word list from a predefined constant file path.
 Master::Master() {
-
     wordlist.read_file(WORDLIST_FILE);
 
     apprenticeList.resize(worldsize - 1);  // Resize apprentice list to match the number of non-master processes.
@@ -56,7 +55,7 @@ void Master::run() {
 // Distributes the word list to all apprentices.
 void Master::distribute_wordlist() {
     int counter{};
-    std::vector<FileData::WordlistData> list{ wordlist.split(apprenticeList.size()) };
+    std::vector<FileData::WordlistData> list{ wordlist.split(static_cast<int>(apprenticeList.size())) };
 
     for (ApprenticeInfo apprentice : apprenticeList) {
         messenger.send_wordlist(apprentice.rank, list.at(counter));
